requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
Django==3.2.5
django-ckeditor==6.1.0
django-js-asset==1.2.2
future==0.18.2
html5lib==1.1
Pillow==8.3.1

sqlparse==0.4.1
webencodings==0.5.1
